partitions code [clang]
COMMENTS:
partitions numpy-like methods and memory handling methods for
spheres into their respective source files
