+++ src/system.h
@@ -1,0 +1,28 @@
+#ifndef GUARD_OPENBDS_SYSTEM_H
+#define GUARD_OPENBDS_SYSTEM_H
+
+#define NUM_SPHERES 256
+
+#endif
+
+/*
+
+OpenBDS								July 19, 2023
+
+source: system.h
+author: @misael-diaz
+
+Synopsis:
+Defines the number of spheres in the system.
+
+Copyright (c) 2023 Misael Diaz-Maldonado
+This file is released under the GNU General Public License as published
+by the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+References:
+[0] A Koenig and B Moo, Accelerated C++ Practical Programming by Example.
+[1] MP Allen and DJ Tildesley, Computer Simulation of Liquids.
+[2] S Kim and S Karrila, Microhydrodynamics: Principles and Selected Applications.
+
+*/
