+++ src/util.h
@@ -1,0 +1,9 @@
+#ifndef GUARD_OPENBDS_UTIL_H
+#define GUARD_OPENBDS_UTIL_H
+
+#include <stdlib.h>
+
+void zeros (size_t const size, double* x);
+void iota (size_t const size, double* x);
+
+#endif
